Dynamic Deep Tanks
Made Deep Tank scan for a cubiod structure of up to 9 blocks in either
direction.
Tank structure changed to use glass walls.
Tank controller and drains must be placed on the outer rim of the bottom
or top layers.
Tank controller must be placed in center of outer rim.
Tank fluid capacity changed to reflect the actual space within the tank.