Make High Ovens more forgiving of structural changes
A structural break no longer resets the heat back to room temperature, but instead cuts it in half immediately and every second thereafter.

This facilitates performing maintenance such as adding an extra duct or drain to the oven while still being able to resume operations in a resonable time after the structure is fixed.