Fix clogging of tank during steam generation
Firstly, we already know that there is enough water in the tank - we just calculated the amount, so no need to check.
Secondly, the drain should be performed before the fill, because we're replacing the same amount - otherwise, the oven clogs, unable to do anything.

Therefore, we do both of these - skip the check, and drain first.